+++ pathman/__init__.py
@@ -1,3 +1,3 @@
-from pathman.path import Path, copy
+from .path import Path
 
-__version__ = "0.1.5"
+__version__ = "0.2.0"